api/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
python-multipart
tensorflow-macos
tensorflow-metal
